Add test. Clear dead VRT extractors
